+++ src/index.ts
@@ -1,5 +1,5 @@
 import VueScreener from './components/VueScreener.vue'
-import VueScreenerPagination from './components/VueScreenerPagination.vue'
+import VueScreenerPagination from './components/pagination/VueScreenerPagination.vue'
 import VueScreenerSearch from './components/VueScreenerSearch.vue'
 import VueScreenerSearchOptions from './components/VueScreenerSearchOptions.vue'
 import VueScreenerTableCell from './components/ui/table/VueScreenerTableCell.vue'
